Change use_ggplot2 to plot_style.
In plotting envelope tests, the argument use_ggplot2 is replaced by the plot_style argument.
The use_ggplot2 can still be used, but the preference is to specify the plot style by the plot_style argument.
The example codes are changed to the new way. The plot style allows "basic" (the previous default), "ggplot2" and "fv" (new).