Replace lexo argument by erl.
Introduce a new, logical argument erl. If TRUE, then the extreme rank length ordering ordering is used both to obtain a p-value and global envelope. Currently still keep lexo,
as a copy for erl. At the same time, a further option erl is given for the ties argument.
If erl = FALSE and this option is chosen, then the erl p-value is calculated, but the global
rank envelope (with p-interval) produced.