Allow different 'theo' in check_curve_set_dimensions.
check_curve_set_dimensions check that the curve_set's in a
given list have similar elements. Previously it was demanded
that all either include or all do not include 'theo'. Now
theo was excluded from the set of the set of components that
need to be either included or not for each curve_set.