Change the first arg of forder to curve_sets.
Allow to give to forder a curve set or a list of curve sets.
In the first case ordering is done by the given measure (and
scaling) for the functions in the curve_set. In the second
case, first the ordering is done by the given measure for each
curve set in the list and, secondly, the extreme rank length
ordering is applied to the first level measures.