Add argument test.equality for graph.fanova.
The test.equality argument can be used to specify that one wants to
test the equality of means by setting test.equality = "mean" (previous functionality), or the equality of variances ("var") or equality of lag s covariances ("cov"). The covariance lag can be specified by the new argument cov.lag.