Re-define fboxplot object.
Let fboxplot object be also a global_envelope object. The basic object therefore is the central region on which the fboxplot is based on. The functional boxplot bounds are further provided as attributes, whisker.lo and whisker.hi. Also the curve_sets is added to attributes for allowing the plotting of outliers.