Change place of converting envelopes to curve_sets.
In calculating the combined global envelope, check at the beginning the reasonability of the curve sets in the list, but do not transform them to curve sets yet. Thus, envelope-objects are kept as envelope-objects and in the first call of individual_central_region the attributes from the envelope-objects are saved. To treat the curves later on, the transformation is done.