Add functionality of one-step combined GEs.
The inner function combined_CR_or_GET now takes in the argument nstep specifying 1 or 2 -step procedure (default to the latter). The function combined_CR_or_GET_1step has the functionality for the one-step procedure for a given list of curve sets. The returned object is a combined_global_envelope.