Allow list of image sets.
The use of global_envelope_test_2d is for an image_set object most likely. However, the functionality exist for a list of image_set objects. Therefore, edit the argument image_set -> image_sets to have it similarly as in global_envelope_test where the first argument is curve_sets.