Update adjusted envelopes/MrkvickaEtal2017=TRUE.
The structure of combined global envelopes has been changed, so the adjusted envelopes need to adjust to the new structure, where, for combined tests, the first level envelopes are given as a list and the second level envelope is found in the attributed level2_ge. This update is for the outdated envelopes from Mrkvicka et al. (2017).