Update of adjusted envelopes.
Update similar to the previous commit, but for the most general case MrkvickaEtal2017=FALSE, type!="rank" (with one function). The additional arguments are saved to the correct places respecting the recent changes in structure of combined global envelopes (the first level envelopes in the main list, no longer in the attribute global_envelope_ls). The second level envelope now for combined tests in attr(..,"level2_ge").