Update GET.necdf.
Use the ecdfmeans.m and ecdfcontrasts.m to calculate means and contrasts. Create a list of curve sets instead of a curve set with long vectors, and call global_envelope_test with nstep = 1, i.e. combining the curves is made within that call. Change the output to be a list of envelopes as well.