Simplify the simulation setup for GET.composite.
In the helper function adj.simulate, change the functionality for the case X is a ppp object. Previously, it was possible to use fitfun and simfun for fitting a model and simulating from it while at the same time using spatstat's envelope for estimating test vectors. Now if fitfun and simfun are provided, then calcfun is used to estimate the test vectors. If not, then for a ppp object X, the tested model if CSR with intensity parameter. This change simplifies the code and allows the use of calcfun for a ppp object (previously not possible). It does not create restrictions, because the user can instead provide X and X.ls as well.