Implement a faster version of frank.fglm.
Add argument fast, which by default is TRUE and then, if no extra arguments are to be passed to lm, a faster algorithm to calculate F-values is used. This is implemented in the inner functions Fvalues, genFvaluesObs and genFvaluesSim. In particular the design matrices are calculated only once and then reused in the permutations as they stay the same over all permutations.