Update helper functions.
Update genCoefmeans.m, genCoefcontrasts.m and genFvaluesLM to take a matrix Y of true or permutated values of the dependent variable and a list of dataframes as their arguments. Thus functionality of creating the data frames was removed from these functions and the data frames are created already earlier and only once.