Update of helper functions.
Replace the function cont_pointwise_hiranks by contrank, which calculates the continuous ranks for a vector corresponding to functions at an argument value (r), and not for functions at all argument values. Add also a new function find_calc_pointwiserank, which define the way to calculate pointwise ranks for different measures and alternatives.