Update checks of object classes.
Replace checks like if(class(curve_sets)[1] == "list") with
more appropriate alternatives checking number of classes and the class,
or using the function inherits.
Also change some checks with is() to inherits for consistency.
Rewrite the orders of checks in the wrapper functions
forder, central_region and global_envelope_test in particular.