Move spatstat to suggests.
Remove direct links to spatstat functions and remove importing spatstat
functions. This is done, because GET can be used without spatstat as well,
spatstat is used for point pattern analysis purposes only.
The usage of spatstat functionalities remains as earlier, if one is
working with spatstat and has installed that in R.