Change GET vignette (.Rnw) to a R.rsp vignette.
This is done to reduce the checking time.
The original GET.Rnw file is included as well as GET.Rnw.orig.
The .tex, and the figures, are obtained by compiling .Rnw, then the line
'% \VignetteEngine{R.rsp::rsp}' is added to the beginning of the .tex file,
and .rsp extension is added.